fix GET order_history response
